(Menu) Fix a segfault in playlist associations

+++ menu/cbs/menu_cbs_left.c
@@ -288,7 +288,7 @@ static int action_left_video_resolution(unsigned type, const char *label,
 static int playlist_association_left(unsigned type, const char *label,
       bool wraparound)
 {
-   size_t i, next, found, current = 0;
+   int i, next, found, current = 0;
    char core_path[PATH_MAX_LENGTH]  = {0};
    char new_playlist_cores[PATH_MAX_LENGTH] = {0};
    global_t *global                 = global_get_ptr();
